Remove padding from Diffable output
When comparing maps where the padding differs between the two sets of keys, the
diff will include keys that were unchanged except for the padding difference.

For example, Arthur Dent appears as a diff due to different padding:
```
{
  Crew: {
-        Arthur Dent:       "Along for the Ride",
-        Zaphod Beeblebrox: "Galactic President",
+        Arthur Dent: "Along for the Ride",
        },
 }
```

This omits padding when `Diffable = true` so that the diff output is consistent
to prevent these spurious diffs.