Switch recursion presentation to labels and references
This is option #3 in

	https://github.com/kylelemons/godebug/pull/17#issuecomment-323546349

Basically while processing current path we detect cycles the same way as it is
done in a85deda10 (Add support for self-referential structures.) and if
a cycle is detected a label and reference elements are correspondingly
inserted, which e.g. for 2-objects cycle below

	type A struct {
	        X int
	        Next *A
	}

        v := &A{1, nil}
        v2 := &A{2, v}
        v.Next = v2

results in output like

	<1>:
	{X:    1,
	 Next: {X:    2,
	        Next: -> <1>}}

Imho labelled output is the most useful way to represent cycles as with
it Compare is always deterministic (in contrast to if there are
differences below current RecursiveContext limit).

The implementation is believed to distinguish cycles from plain cases
where one object was simply referenced twice from different places, and
a test is added for this case.